packages/core/pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = "Shared core utilities for AI agent framework"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^1.0.0"
langchain = "^0.1.0"
langchain-openai = "^0.0.5"
redis = "^5.0.0"
sqlmodel = "^0.0.14"
psycopg2-binary = "^2.9.9"
pgvector = "^0.2.0"
pinecone-client = "^3.0.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.11.0"
mypy = "^1.7.0"
ruff = "^0.1.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

